atualizar

+++ lavanderia-controle.php
@@ -10,32 +10,51 @@ function buscarCliente($conexao,$login){
 }
 
 function inserirCliente($conexao,$nome,$email,$telefone,$login,$senha){
-    $sql="INSERT INTO cliente(nome,email,telefone,login,senha) VALUES('$nome','$email','$telefone','$login','$senha')";
+    $sql="INSERT INTO cliente(nome,email,telefone,login,senha,tipo) VALUES('$nome','$email','$telefone','$login','$senha',$tipo)";
     mysqli_query($conexao,$sql) or die (mysqli_error($conexao));
 }
 
-function alterarCliente($conexao,$id,$nome,$email,$telefone,$login,$senha){
-    $sql="UPDATE cliente SET nome='$nome',email='$email',telefone='$telefone',login='$login',senha='$senha', WHERE id=$id";
+function alterarCliente($conexao,$id,$nome,$email,$telefone,$login,$senha,$tipo){
+    if($tipo=="admin"){
+    $sql="UPDATE cliente SET nome='$nome',email='$email',telefone='$telefone',login='$login',senha='$senha',tipo='$tipo' WHERE id=$id";
+    }else{
+    $sql="UPDATE cliente SET nome='$nome',email='$email',telefone='$telefone',login='$login',senha='$senha',tipo='$tipo' WHERE id=$id AND cliente=$id";
+    }
 
     mysqli_query($conexao,$sql) or die( mysqli_error($conexao));
 }
-function atualizaSenha($conexao,$id,$login,$senha){
-$sql="UPDATE cliente SET login='$login',senha='$senha' WHERE id=$id";
+
+function atualizaSenha($conexao,$id,$login,$senha,$tipo){
+    if($tipo=="admin"){
+    $sql="UPDATE cliente SET login='$login',senha='$senha',tipo='$tipo' WHERE id=$id";
+    }else{
+    $sql="UPDATE cliente SET login='$login',senha='$senha',tipo='$tipo' WHERE id=$id AND cliente=$id";
+    }
 mysqli_query($conexao,$sql) or  die (mysqli_error($conexao));
 }
 
-function excluirCliente($conexao,$id){
-    
+function excluirCliente($conexao,$id,$tipo){
+    if($tipo){
     $sql="DELETE FROM cliente WHERE id=$id"; 
-    
+    }else{
+        $sql="DELETE FROM cliente WHERE id=$id AND cliente=$id"; 
+
+    }
     mysqli_query($conexao,$sql) or  die(mysqli_error($conexao));
 }
+
 function listaCliente($conexao){
+    if($ipo=="admin"){
     $sql="SELECT id, nome,email,telefone,login,senha FROM cliente";
+    }else{
+        $sql="SELECT id, nome,email,telefone,login,senha FROM cliente WHERE =id";
+
+    }
 
    $resultado= mysqli_query($conexao,$sql) or  die (mysqli_error($conexao));
    return mysqli_fetch_all($resultado,MYSQLI_ASSOC);
 }
+
 function listarUmCliente($conexao,$id){
 $sql="SELECT * FROM cliente WHERE id=$id";
 $resultado= mysqli_query($conexao,$sql) or  die(mysqli_error($conexao));
